third_party/WebKit/LayoutTests/external/wpt/webvtt/parsing/file-parsing/tests/settings-line-expected.txt

This is a testharness.js-based test.
FAIL settings, line assert_equals: Failed with cue 33 expected (string) "auto" but got (number) Infinity
Harness: the test ran to completion.
